qemuboot: move QB_MEM to default in qemuboot from core-image-sato
Previously, to satisfy opengl's requirement for at least 512MB
of memory, QB_MEM was overriden by core-image-sato. This made it
impossible for machine configurations to specify their own values,
so this clause was moved to qemuboot instead and rewritten as a
default to allow other recipes to state their own values if
needed.