qemuboot: change QB_MEM default strengths
Previously, to satisfy opengl's requirement for at least 512MB
of memory, QB_MEM was overriden by core-image-sato. This made it
impossible for machine configurations to specify their own values,
so this was changed to a default asssignment. To accomodate this,
qemuboot's default assignment of QB_MEM was changed to the week
variant.

Change-Id: I83d7fff3ff00dcfb1300bf5ce46cffdd125facd5
Signed-off-by: Brett Warren <brett.warren@arm.com>